+++ components/Content/index.module.scss
@@ -1,158 +1,40 @@
 .Content {
   display: flex;
   flex-direction: column;
-  margin: auto 15%;
-  width: 100vw !important;
+  margin: 15px 100px 100px 200px;
+  width: 85 vw;
+  overflow: auto;
 
   &__wrapper {
-    display: flex;
+    display: grid;
     padding: 32px;
     gap: 40px;
-    overflow-x: none;
-    border: 1px solid red;
-    width: 500px !important;
-    position: relative;
-
-    &__tabs {
-      display: flex;
-      justify-content: space-evenly;
-      margin-left: 10px;
-
-      @media (max-width: 768px) {
-        flex-direction: row;
-        justify-content: space-between;
-        // border: 1px solid red;
-        width: 100%;
-      }
-      &__left {
-        // margin-right: 350px;
-        h2 {
-          font-weight: bold;
-        }
-        p {
-          color: #718096;
-        }
-
-        flex-basis: 50%;
-      }
-      &__right {
-        // margin-left: 280px;
-        flex-basis: 50%;
-        &__in {
-          margin-bottom: 20px;
-          font-size: medium;
-          font-weight: bolder;
-          &__one {
-            text-align: center;
-            padding: 15px 25px;
-            border: 1px solid rgb(203, 240, 240);
-            border-radius: 10px;
-          }
-          &__two {
-            background: rgb(227, 226, 226);
-            text-align: center;
-            padding: 15px 25px;
-            // border: 4px solid black;
-          }
-        }
-        &__in {
-          display: flex;
-          gap: 4px;
-        }
-        &__dot {
-          display: flex;
-          h3 {
-            margin: -10px 15px;
-          }
-          &__in {
-            width: 10px;
-            height: 10px;
-            border-radius: 50%;
-            background-color: rgba(75, 192, 192, 0.4);
-          }
-          &__out {
-            width: 10px;
-            height: 10px;
-            border-radius: 50%;
-            background-color: rgba(75, 192, 192, 1);
-          }
-        }
-      }
-    }
+    width: auto;
+    // overflow-x: none;
   }
   .charts {
     display: flex;
     flex-direction: column;
-    justify-content: space-evenly;
-
-    &__bar {
-      margin-left: 20px;
-    }
     &__sales {
       display: flex;
       justify-content: space-evenly;
-      margin-top: 40px;
+      margin-top: 40px 0px 0px 0px;
 
-      &__invoices {
+      &__progressbar {
         display: flex;
-        margin-left: 60px;
-        &__right {
-        }
-        &__leftside {
-          display: flex;
-          flex-direction: column;
-          margin-top: 50px;
-          &__data {
-            display: flex;
-            h2 {
-              margin-top: 0;
-            }
-            p {
-              margin: 30px -50px;
-              color: #718096;
-              font-size: large;
-            }
-            &__dot {
-              margin: 10px 10px;
-              width: 10px;
-              height: 10px;
-              border-radius: 50%;
-            }
-          }
-        }
-      }
-      &__expenses {
-        &__amount {
-          display: flex;
-          width: 400px;
-          &__left {
-            align-self: flex-start;
-          }
-          &__right {
-            margin-left: 150px;
-          }
-        }
-        &__progress {
-          width: 400px;
-          height: 40px;
-          background: #ddd;
-          border-radius: 15px;
-          &__bar {
-            background: rgb(153, 201, 195);
-            height: 100%;
-            width: 20%;
-            border-top-left-radius: 10px;
-            border-bottom-left-radius: 10px;
-          }
-        }
+        flex-direction: column;
+        // margin-top: 50px;
       }
     }
   }
 }
+
 @media only screen and (max-width: 768px) {
   .Content {
-    // margin-left: 70px;
-    // overflow-x: auto;
+    margin-left: 70px;
+    overflow-x: auto;
+    // display: flex;
+    // flex-direction: column;
 
     &__wrapper {
       display: grid;
@@ -165,32 +47,22 @@
       }
     }
     .charts {
-      display: flexbox;
+      // display: flexbox;
       flex-direction: column;
-      justify-content: center;
+      justify-content: space-between;
       align-items: center;
       margin-left: auto;
 
-      &__bar > h2 {
-        text-align: center;
-        font-size: 22px;
-        margin-bottom: 20px;
-      }
       &__sales > h2 {
         text-align: center;
         font-size: 22px;
         margin-bottom: 20px;
       }
-      &__bar {
-        margin-top: 30px;
-        margin-left: -500px;
-        width: 200px;
-        height: 300px;
-      }
+
       &__sales {
+        margin-top: 50px;
         display: flex;
         flex-direction: column;
-        margin-left: 10px;
         width: 300px;
         height: 300px;
       }
