+++ components/Header/index.module.scss
@@ -3,19 +3,17 @@
   flex-wrap: wrap;
   height: 70px;
   background-color: white;
-
+  position: relative;
+  margin-top: 20px;
   &__wrapper {
     display: flex;
-    margin-left: 200px;
-    // justify-content: space-evenly;
-    // align-items: center;
-
+    padding-left: 40px;
     @media (max-width: 768px) {
       margin-left: 300px;
     }
     &__right {
       display: flex;
-      margin: 15px 60px 0px 480px;
+      margin-left: 600px;
       &__search {
         // margin: 10px 20px 0 400px;
       }
@@ -25,8 +23,12 @@
         right: 20px;
         width: 40px;
         height: 40px;
-        & img {
-          border-radius: 50%;
+        color: grey;
+        border-radius: 50%;
+
+        background: rgb(247, 247, 247);
+        &_image {
+          margin: 7px !important;
           cursor: pointer;
         }
       }
@@ -45,13 +47,14 @@
       //   }
     }
     &__title {
-      margin-left: 40px;
-    }
-    &__title > h2 {
-      color: rgb(0, 0, 0);
-    }
-    &__title > p {
-      color: rgb(0, 0, 0);
+      // margin-top: -20px 10px 0px 20px;
+      &__logo {
+        // img {
+        margin-top: 5px;
+        width: 60px !important;
+        height: 50px;
+        // }
+      }
     }
   }
 }
