+++ components/charts/BarComponent/index.module.scss
@@ -1,0 +1,10 @@
+// .bar_container__graph__chart {
+//   width: 800px;
+//   height: 250px;
+// }
+// @media only screen and (max-width: 768px) {
+//   .bar_container__graph__chart {
+//     width: 300px;
+//     height: 200px;
+//   }
+// }
